chore: moves cleanup logic to single file
